data/scripts/data_acquisition.py
# data/scripts/data_acquisition.py

import os
import xarray as xr
import requests
import argparse
import yaml

def download_cmip6_data(config):
    # Placeholder function to demonstrate data download
    print("Downloading CMIP6 data...")
    # Actual implementation would involve ESGF APIs and user authentication

def download_era5_data(config):
    # Placeholder function to demonstrate data download
    print("Downloading ERA5 data...")
    # Actual implementation would involve Copernicus Climate Data Store API

def main(config_path):
    # Load configuration
    with open(config_path) as f:
        config = yaml.safe_load(f)
    
    # Create data directories if they don't exist
    os.makedirs(config['data_paths']['raw_cmip6'], exist_ok=True)
    os.makedirs(config['data_paths']['raw_era5'], exist_ok=True)
    
    # Download data
    download_cmip6_data(config)
    download_era5_data(config)
    
    print("Data acquisition complete.")

if __name__ == "__main__":
    parser = argparse.ArgumentParser(description='Data Acquisition Script')
    parser.add_argument('--config', type=str, help='Path to the config file', default='configs/data_config.yaml')
    args = parser.parse_args()
    main(args.config)
