requirements.txt
numpy>=1.18.0
xarray>=0.15.0
netCDF4>=1.5.3
pandas>=1.0.0
scikit-learn>=0.22.0
matplotlib>=3.1.0
tensorflow>=2.0.0
keras>=2.3.0
xesmf>=0.3.0
shap>=0.35.0
lime>=0.2.0
cartopy>=0.17.0
dask>=2.9.0
opencv-python>=4.2.0.32
PyYAML>=5.3
